Update sass compile script to no longer assume /sass exists in /css dirs

+++ scripts/compile_css.sh
@@ -71,20 +71,24 @@ function main () {
                 # Check if watch command was set. Annoyingly, this changes sass syntax.
                 if [[ $watch == "--update" ]]
                 then
-                    # Loop through all files in sass subfolder.
-                    for file in $dir/sass/*
-                    do
-                        # Check that file follows sass compilation file naming convention.
-                        if [[ $file != *"/css/sass/_"*".scss" ]]
-                        then
-                            # Add file to list of compile locations.
-                            filename=$(basename "${file%.*}")
-                            css_directories+=("$file:$dir/$filename.css")
+                    # Check that sass directory exists.
+                    if [[ -d $dir/sass ]]
+                    then
+                        # Loop through all files in sass subfolder.
+                        for file in $dir/sass/*
+                        do
+                            # Check that file follows sass compilation file naming convention.
+                            if [[ $file != *"/css/sass/_"*".scss" ]]
+                            then
+                                # Add file to list of compile locations.
+                                filename=$(basename "${file%.*}")
+                                css_directories+=("$file:$dir/$filename.css")
 
-                            # Remove old file before compiling, if present.
-                            rm -f "$dir/$filename.css" "$dir/$filename.css.map"
-                        fi
-                    done
+                                # Remove old file before compiling, if present.
+                                rm -f "$dir/$filename.css" "$dir/$filename.css.map"
+                            fi
+                        done
+                    fi
                 else
                     # Add directory to list of compile locations.
                     css_directories+=("$dir/sass:$dir")
